(#1591) ensure correct creds are used for jetstream check
